extracted file upload functionality
